# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trosc API Documentation",
    "version": "1.0.0",
    "description": "Official API documentation for the Trosc platform.",
    "contact": {
      "name": "Basem Esam",
      "url": "https://github.com/basem3sam"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/v1",
      "description": "development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CourseResponse": {
        "type": "object",
        "description": "Standard response format for course operations",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "course": {
                "$ref": "#/components/schemas/Course"
              }
            }
          }
        }
      },
...
      "Resource": {
        "type": "object",
        "description": "Learning resource attached to a session",
        "required": [
          "title",
          "url"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the resource",
            "example": "JavaScript Arrays Cheat Sheet"
          },
          "url": {
            "type": "string",
            "description": "URL to access the resource",
            "format": "uri",
            "example": "https://example.com/arrays-cheatsheet.pdf"
          }
        },
        "example": {
          "title": "Practice Exercises PDF",
          "url": "https://example.com/exercises.pdf"
        }
      },
...
      "SessionCreate": {
        "type": "object",
        "description": "Data required to create a new session",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "JavaScript Functions Deep Dive"
          },
          "description": {
            "type": "string",
            "example": "Learn about function declarations, expressions, arrow functions, and closures"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
          },
          "isStandalone": {
            "type": "boolean",
            "example": false
          },
          "duration": {
            "type": "integer",
            "example": 90
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "intermediate"
          },
          "coverImage": {
            "type": "string",
            "example": "functions-session-cover.jpg"
          },
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "published": {
            "type": "boolean",
            "example": true
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-20T14:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-20T15:30:00.000Z"
          }
        }
      },
      "SessionUpdate": {
        "type": "object",
        "description": "Data that can be updated for a session",
        "properties": {
          "title": {
            "type": "string",
            "example": "Updated Session Title"
          },
          "description": {
            "type": "string",
            "example": "Updated session description"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "example": "https://drive.google.com/file/d/abc123"
          },
          "instructor": {
            "type": "string",
            "example": "507f1f77bcf86cd799439012"
          },
          "tracks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "507f1f77bcf86cd799439021"
            ]
          },
          "isStandalone": {
            "type": "boolean",
            "example": true
          },
          "duration": {
            "type": "integer",
            "example": 120
          },
          "level": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "example": "advanced"
          },
          "coverImage": {
            "type": "string",
            "example": "new-cover-image.jpg"
          },
          "resources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          },
          "published": {
            "type": "boolean",
            "example": false
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-25T10:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-10-25T12:00:00.000Z"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "description": "Response format for a single session",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "session": {
                "$ref": "#/components/schemas/Session"
              }
            }
          }
        }
      },
      "SessionsResponse": {
        "type": "object",
        "description": "Response format for multiple sessions",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "results": {
            "type": "integer",
            "example": 8
          },
          "total": {
            "type": "integer",
            "example": 50
          },
          "data": {
            "type": "object",
            "properties": {
              "sessions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        }
      },
...
      "TrackResponse": {
        "type": "object",
        "description": "Standard response format for track operations",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "data": {
            "type": "object",
            "properties": {
              "track": {
                "$ref": "#/components/schemas/Track"
              }
            }
          }
        }
      },
...
      "StandardError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "fail"
          },
          "message": {
            "type": "string",
            "example": "Error description here"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/courses/{courseId}/sessions/{sessionId}": {
      "patch": {
        "operationId": "addSessionToCourse",
        "summary": "Add a session to a course",
        "description": "Associate a session with a course (admin and instructors only). Session can be standalone or from any track.",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439041"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439031"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session added to course successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Session already in course or belongs to different track"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Course or session not found"
          }
        }
      },
      "delete": {
        "operationId": "removeSessionFromCourse",
        "summary": "Remove a session from a course",
        "description": "Remove session association from a course (admin and instructors only)",
        "tags": [
          "Courses"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439041"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439031"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session removed from course successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Session not found in course"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Course or session not found"
          }
        }
      }
    },
...
    "/sessions": {
      "post": {
        "operationId": "createSession",
        "summary": "Create a new learning session",
        "description": "Create a new session (admin and instructors only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "get": {
        "operationId": "getAllSessions",
        "summary": "Get all sessions",
        "description": "Retrieve all sessions with filtering and pagination",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "level",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "beginner",
                "intermediate",
                "advanced"
              ]
            }
          },
          {
            "name": "published",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of sessions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/sessions/{id}": {
      "get": {
        "operationId": "getSessionById",
        "summary": "Get a specific session by ID",
        "description": "Retrieve detailed information about a session",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "operationId": "updateSessionById",
        "summary": "Update a session",
        "description": "Update session information (admin and instructors only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteSessionById",
        "summary": "Delete a session",
        "description": "Permanently delete a session (admin only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Session deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/sessions/{id}/students": {
      "post": {
        "operationId": "addStudentToSession",
        "summary": "Add student to session",
        "description": "Enroll a student in a session (admin and instructors only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "studentId"
                ],
                "properties": {
                  "studentId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student added to session successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Student already enrolled in this session"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/students/{studentId}": {
      "delete": {
        "operationId": "removeStudentFromSession",
        "summary": "Remove student from session",
        "description": "Remove a student from a session (admin and instructors only)",
        "tags": [
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439031"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "description": "MongoDB ID of the student to remove",
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439012"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Student removed from session successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Student is not enrolled in this session"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Session or student not found"
          }
        }
      }
    },
...
    "/tracks/{trackId}/sessions/{sessionId}": {
      "patch": {
        "operationId": "addSessionToTrack",
        "summary": "Add a session to a track",
        "description": "Associate a session with a track (admin and instructors only)",
        "tags": [
          "Tracks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439021"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439031"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session added to track successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Session already exists in track or invalid operation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Track or session not found"
          }
        }
      },
      "delete": {
        "operationId": "removeSessionFromTrack",
        "summary": "Remove a session from a track",
        "description": "Remove session association from a track (admin and instructors only)",
        "tags": [
          "Tracks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439021"
            }
          },
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "507f1f77bcf86cd799439031"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session removed from track successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackResponse"
                }
              }
            }
          },
          "400": {
            "description": "Session not found in track"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "description": "Track or session not found"
          }
        }
      }
    },
...
  },
...
}